requirements.txt
feedparser==6.0.11
httpx==0.27.0
anthropic==0.34.0
python-dotenv==1.0.1
